+++ apps/management-app/src/components/Common/Layout/PageLayout.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@mui/material';
 import { Outlet } from 'react-router-dom';
-import Header from './Common/Header';
+import Header from './Header';
 
 function PageLayout() {
   return (
